Refactor. Add helper
